python-code-backend/requirements.txt
PyMuPDF
openai
pinecone
python-dotenv
tqdm
selenium
requests
beautifulsoup4
lxml
urllib3
webdriver-manager
